append B to contentlength only when its non zero
