+++ cosmos-server/src/main/scala/com/mesosphere/cosmos/package.scala
@@ -6,6 +6,12 @@ import com.mesosphere.error.ResultOps
 import com.netaporter.uri.Uri
 import com.twitter.app.GlobalFlag
 import com.twitter.conversions.storage._
+import com.twitter.finagle.http.Fields
+import com.twitter.finagle.http.Request
+import com.twitter.finagle.http.Response
+import com.twitter.finagle.http.filter.CommonLogFormatter
+import com.twitter.finagle.http.filter.LoggingFilter
+import com.twitter.util.Duration
 import com.twitter.util.ScheduledThreadPoolTimer
 import com.twitter.util.StorageUnit
 import com.twitter.util.Timer
@@ -14,12 +20,54 @@ import java.nio.file.Path
 
 package object cosmos {
   implicit val globalTimer: Timer = new ScheduledThreadPoolTimer()
+
+  object CustomLoggingFilter extends LoggingFilter[Request](
+    log = com.twitter.logging.Logger("access"),
+    formatter = new CommonLogFormatter {
+      override def format(request: Request, response: Response, responseTime: Duration): String = {
+        val remoteAddr = request.remoteAddress.getHostAddress
+
+        val contentLength = response.length
+        val contentLengthStr = if (contentLength > 0) s"${contentLength.toString}B" else "-"
+
+        val builder = new StringBuilder
+        builder.append(remoteAddr)
+        builder.append(" - \"")
+        builder.append(escape(request.method.toString))
+        builder.append(' ')
+        builder.append(escape(request.uri))
+        builder.append(' ')
+        builder.append(escape(request.version.toString))
+        builder.append("\" ")
+        builder.append(response.statusCode.toString)
+        builder.append(' ')
+        builder.append(contentLengthStr)
+        builder.append(' ')
+        builder.append(responseTime.inMillis)
+        builder.append("ms \"")
+        builder.append(escape(request.userAgent.getOrElse("-")))
+        builder.append('"')
+
+        if (response.statusCode / 100 != 2) {
+          val headersMap = request.headerMap
+          headersMap.get(Fields.Authorization) match {
+            case Some(_) => headersMap.put(Fields.Authorization, "********")
+            case None => ()
+          }
+          builder.append(s" Headers : (${headersMap.map(_.productIterator.mkString(":")).mkString(", ")})")
+        }
+
+        builder.toString
+      }
+    }
+  )
 }
 
 /* A flag's name is the fully-qualified classname. GlobalFlag doesn't support package object. We
  * must instead use regular package declarations
  */
 package cosmos {
+
   // scalastyle:off object.name
   object dcosUri extends GlobalFlag[Uri](
     s"The URI where the DCOS Admin Router is located. If this flag is set, " +
@@ -75,4 +123,5 @@ package cosmos {
     "Maximum size for the proxy endpoint service while fetching a resource"
   )
   // scalastyle:on object.name
+
 }
